BACK-1606 Reduce save balance history memory footprint (#254)
* Reduce save balance history memory footprint

On huge account first synchronization, the amount of balance histories
may be huge ( > 42k ).
In order to limit the footprint, the stream of balance histories is
sliced into chunks. The size of the chunk is between 100 and 1000 items
depending of the number of concurrent batch allowed in configuration
(db.batch-concurrency)

The default db.batch-concurrency is by default of 5 concurrent db
transactions. It may be overridden with the env variable
BITCOIN_INTERPRETER_TWEAK_DB_BATCH_CONCURRENCY

 * BatchConcurrency must not be negative